feat : 자유게시판 좋아요 기능

+++ src/main/java/com/example/coffies_vol_02/Like/service/LikeService.java
@@ -9,8 +9,6 @@ import com.example.coffies_vol_02.Config.Exception.Handler.CustomExceptionHandle
 import com.example.coffies_vol_02.Like.domain.Like;
 import com.example.coffies_vol_02.Like.repository.LikeRepository;
 import com.example.coffies_vol_02.Member.domain.Member;
-import com.example.coffies_vol_02.Member.repository.MemberRepository;
-import com.example.coffies_vol_02.Place.domain.Place;
 import com.example.coffies_vol_02.Place.repository.PlaceRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -33,7 +31,7 @@ public class LikeService {
     */
     @Transactional
     public boolean hasLikeBoard(Board board, Member member){
-        return likeRepository.findByMemberAndBoard(member,board).isEmpty();
+        return likeRepository.findByMemberAndBoard(member,board).isPresent();
     }
 
     /*
@@ -43,8 +41,11 @@ public class LikeService {
         Optional<Board>detail = Optional.ofNullable(boardRepository.findById(boardId).orElseThrow(() -> new CustomExceptionHandler(ERRORCODE.BOARD_NOT_LIST)));
 
         //좋아요 중복체크를 거친 뒤에 중복되지 않으면 카운트,
-        if(hasLikeBoard(detail.get(),member)){
+        if(hasLikeBoard(detail.get(),member) == false){
             likeRepository.save(new Like(member, detail.get()));
+        }else if(hasLikeBoard(detail.get(),member)== true){
+            cancelLike(detail.get().getId(),member);
+            return LikeCancel;
         }
         return LikeSuccess;
     }
@@ -57,8 +58,12 @@ public class LikeService {
 
         Optional<Like> like = Optional.ofNullable(likeRepository.findByMemberAndBoard(member, detail.get()).orElseThrow(() -> new CustomExceptionHandler(ERRORCODE.LIKE_NOT_FOUND)));
 
-        likeRepository.delete(like.get());
-
+        if(like.isPresent()){
+            likeRepository.delete(like.get());
+        }else{
+            createBoardLike(boardId,member);
+            return LikeSuccess;
+        }
         return LikeCancel;
     }
 
@@ -78,6 +83,7 @@ public class LikeService {
         }
         return resultData;
     }
+
     @Transactional
     public boolean hasLikeComment(Member member, Comment comment){
         return likeRepository.findByMemberAndComment(member,comment).isEmpty();
